Fix bracketed-expression bug
This fixes a bug which meant that expressions in brackets ending in an xpath,
but withouth a leading space before the closing bracket, would fail to parse.

E.g.:

 Would fail: (1 * /x/path)
 Would pass: (1 * /x/path )

This commit also takes the opportunity to tighten up the conditions checking in
integration.spec.js to only use regexes where it is acutally nexcessary, and in
other cases use more specific data types.