Convert date extensions to plain functions
This came up in enketo-core#791, where certain tests were depending on the Date.prototype monkeypatch side-effect before it had been performed. The intent in this change is to deprecate the prototype pollution by providing the same functionality as plain functions. Tests were adapted from medic/cht-core which appears to test a slightly different version of the same code.