Fix version in openrosa-xpath-evaluator lock file
Note: this should have been run prior to release. It was set automatically when validating `npm install` in the package directory.

More importantly, at present `npm install` will only succeed on aarch64 (i.e. on a Mac with M1 or later processor) with the `--ignore-scripts` flag. This is also the case outside of this monorepo. It’s mentioned here to track the issue as I continue with monorepo setup.