Remove built docs
It's not clear why those were included. It seems that if someone is developing EE, they'll find the inline docs more useful. As far as I know, none of this is intended to be a public interface. The current maintainers do not consider any implementation to be a public interface.