Resolve custom widgets in enketo-express config
This follows roughly the same logic used to resolve first party widgets (i.e. from enketo-core). The substantive difference is that it defers to `require.resolve` rather than any assumptions about where in the file system the widget will be installed—this is a minor safeguard in case, for example, a widget install isn’t hoisted to the monorepo root.

This change also adds a console warning both during build and in the generated widget import module, for any widgets (first party or custom) which fail to resolve.