feat(audioRecording): Implement audio recording and playback functionality DEV-210 (#1413)
### 📣 Summary
This PR implements the actual functionality of capturing audio data for
recording and picking a file. The playback of the recorded or selected
file was also implemented.

### 📖 Description
⚠️ This PR is to be merged into the feat-audio-recording feature branch.
The audio recording feature will only be available in main when it's
fully implemented.

This PR adds to the audio recording widget the functionalities:
- Record audio
- Pause and resume audio recording
- Upload an audio file
- Playback the recorded or uploaded file
- Download the recorded or uploaded file
- Preview audio waveform for feedback during recording

Also:
- The second field form `audio-recording.xml` form was set as required
- The validation for the required fiel will be implemented in the next
task. For now, the field will only be validated when a file is pick, but
not when the audio is recorded.
- Confirmation dialog when deleting the audio was added

### 👀 Preview steps

1. ℹ️ Have the form server and Enketo express running (use `yarn watch`)
2. Open the form preview to check the audio-recording.xml form
(http://localhost:8005/preview?xform=http://localhost:3000/audio-recording.xml)
3. 🟢 The audio recording widget should load properly in the form for
interaction
4. Also check the all-widgets.xml form
(http://localhost:8005/preview?xform=http://localhost:3000/all-widgets.xml)
5. 🟢 The audio recording widget should load properly in the form for
interaction

---------

Co-authored-by: Akuukis <akuukis@kalvis.lv>