feat(audioRecording): validate form content and attach audio to submission DEV-211 (#1415)
### 📣 Summary
This PR wraps the development of the audio recording widget by
implementing:
- Sending of the submission with the recorded audio to the server 
- Editing submission capabilities for the audio recording widget
- Correct validation of audio recording widget data

### 💭 Notes
Enketo uses the input elements of the form to control the data to be
sent as the submission, but for files being generated (e.g. the
draw-widget) a `data-cache` attribute is also used to store the data for
the actual file, while the input value is only used to keep track of the
file name.
When editing a submission, a `data-loaded-file-name` attribute is set in
the `input` element so that we know it has a value set.

I have verified this PR works by manually testing (see CONTRIBUTING.md):

- [x] Online form submission
- [x] Offline form submission
- [x] Saving offline drafts
- [x] Loading offline drafts
- [x] Editing submissions
- [x] Form preview
- [ ] None of the above

### 👀 Preview steps
Follow [the
instructions](https://github.com/enketo/enketo/blob/1da7b25bd8dc3fdc48ba9f31190887d036408360/packages/enketo-express/tutorials/developing-with-kpi.md)
to get a local Enketo instance connected to a KoboToolbox environment

1. ℹ️ have an account and a project
2. Create a form with audio questions (Or you can use
[audioRecordingForm.xlsx](https://github.com/user-attachments/files/21514085/audioRecordingForm.xlsx)
to create a new form)
3. Deploy the form
4. Open the form for Preview
5. Validate form data
6. Required field should be properly validated
7. Open the form for submitting data
8. Submit recorded audio
9. Submission should appear properly in the project's Data table
10. Edit a submission (from KoboToolbox)
11. Submission data should be properly edited