feat(widgetValidation): implement custom widget validation DEV-1036 (#1444)
### 📣 Summary
This PR implements a new custom widget validation, called right before
submitting the form and taking precedence over `required` and
`constraints`. It is available for widgets through the overwriting of
the `validate` function.

### 💭 Notes
This PR only adds the validation functionality, but does not apply the
validation in any widget.

### 👀 Preview steps

There's no user-facing changes. If needed, this can be tested in code in
the following way:

- Pull the PR
- Pick a widget (e.g.: audio.js)
- Add the `validate` function
- Add any logic and use `setValidationError(error message)` to set the
error in the widget: e.g.:
   ```javascript
   class AudioWidget extends Widget {
    ...
    validate() {
        if (!this.tries || this.tries < 3) {
            this.tries = this.tries ? this.tries + 1 : 1;
            this.setValidationError(
                `Not enough tries. Only ${this.tries} tries. Try again.`
            );
            return;
        }
        this.setValidationError(null);
    }
    ...
   }
   ```
- Call `setValidationError(null)` to clear the error
- Load a form containing the widget (e.g:
[multipageAudio.xlsx](https://github.com/user-attachments/files/22660413/multipageAudio.xlsx))

- Try to submit the form and it should be blocked by the added logic if
an error was set