fix(security): sanitize text values in print view DEV-637 (#1460)
### 📣 Summary

Tjhis PR fixes a stored XSS vulnerability in form print preview that
allowed malicious default values to execute as JavaScript.


### 📖 Description

This PR addresses a security vulnerability (Cure53 ODK-01-013 WP2) where
form values containing HTML/JavaScript could execute when printing a
form preview. The vulnerability was present when a form contained
malicious code in input values. Those values could be default and even
hidden, allowing for malicious code execution in forms when printing.


### 💭 Notes

When printing, widget values were being set as `innerHTML` without any
sanitization for some newly created html elements to generate a full
printable HTML.
Sanitization was added for the input values in the created elements. 
A test form for this issue was added to the forms so the vulnerability
can be tested and validated.


### 👀 Preview steps
⚠️ **main branch will not have the test form for a switching branches.
So the new `XSS-test.xml` form needs to be copied into main for the
reproduction of the issue**

1. run enketo locally with `yarn watch`, so either enketo-express and
the forms server are running (KPI is not needed)
2. open
`http://localhost:8005/preview?form=http://localhost:3000/XSS-test.xml`
3. click the print button in the top right of the form
4. cancel the printing
5. 🔴 [on main] the malicious code is executed and an alert appears in
the browser
6. 🟢 [on PR] nothing happens (no malicious code is executed)